# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/locacao": {
      "post": {
        "operationId": "LocacaoController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLocacaoDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Locacao"
        ]
      },
      "get": {
        "operationId": "LocacaoController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Locacao"
        ]
      }
    },
    "/locacao/{id}": {
      "get": {
        "operationId": "LocacaoController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Locacao"
        ]
      },
      "patch": {
        "operationId": "LocacaoController_updateLocacao",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLocacaoDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Locacao"
        ]
      },
      "delete": {
        "operationId": "LocacaoController_deleteLocacaoById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Locacao"
        ]
      }
    },
    "/locacao/checkins/turno": {
      "get": {
        "operationId": "LocacaoController_getCheckinsTurno",
        "parameters": [
          {
            "name": "usuario_id",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "motel_id",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "horas",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "posicao",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Locacao"
        ]
      }
    },
    "/locacao/{id}/checkout": {
      "patch": {
        "operationId": "LocacaoController_checkout",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "desconto",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Locacao"
        ]
      }
    },
...
    "/pessoa/{id}": {
      "get": {
        "operationId": "PessoaController_findOnePessoa",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Pessoa"
        ]
      },
      "patch": {
        "operationId": "PessoaController_updatePessoa",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePessoaDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Pessoa"
        ]
      },
      "delete": {
        "operationId": "PessoaController_removePessoa",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Pessoa"
        ]
      }
    },
    "/pessoatipo": {
      "post": {
        "operationId": "PessoaTipoController_createPessoaTipo",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePessoaTipoDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "PessoaTipo"
        ]
      },
      "get": {
        "operationId": "PessoaTipoController_findAllPessoaTipos",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "PessoaTipo"
        ]
      }
    },
    "/pessoatipo/{id}": {
      "get": {
        "operationId": "PessoaTipoController_findOnePessoaTipo",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "PessoaTipo"
        ]
      },
      "patch": {
        "operationId": "PessoaTipoController_updatePessoaTipo",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePessoaTipoDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "PessoaTipo"
        ]
      },
      "delete": {
        "operationId": "PessoaTipoController_removePessoaTipo",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "PessoaTipo"
        ]
      }
    },
...
    "/registroponto": {
      "post": {
        "operationId": "RegistroPontoController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRegistroPontoDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "RegistroPonto"
        ]
      },
      "get": {
        "operationId": "RegistroPontoController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "RegistroPonto"
        ]
      }
    },
    "/registroponto/{id}": {
      "get": {
        "operationId": "RegistroPontoController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "RegistroPonto"
        ]
      },
      "patch": {
        "operationId": "RegistroPontoController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRegistroPontoDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "RegistroPonto"
        ]
      },
      "delete": {
        "operationId": "RegistroPontoController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "RegistroPonto"
        ]
      }
    },
...
  },
  "info": {
    "title": "Luvr - Gestão Inteligente",
    "description": "Escola de T.I - software de gestão de motel.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "https://api.luvr.com.br"
    },
    {
      "url": "http://localhost:3000"
    }
  ],
  "components": {
    "schemas": {
...
      "CreateLocacaoDto": {
        "type": "object",
        "properties": {
          "locacao_totalItens": {
            "type": "number",
            "example": 129
          },
          "locacao_totalQuarto": {
            "type": "number",
            "example": 240
          },
          "locacao_totalDesconto": {
            "type": "number",
            "example": 20
          },
          "locacao_totalLocacao": {
            "type": "number",
            "example": 349
          },
          "motel_id": {
            "type": "number",
            "example": 1
          },
          "quarto_id": {
            "type": "number",
            "example": 10
          },
          "pessoa_id": {
            "type": "number",
            "example": 5
          },
          "pagamentoforma_id": {
            "type": "number",
            "example": 2
          },
          "usuario_id": {
            "type": "number",
            "example": 3
          },
          "locacaoPosicao_id": {
            "type": "number",
            "example": 1
          },
          "locacaoTipo_id": {
            "type": "number",
            "example": 1
          }
        },
        "required": [
          "locacao_totalItens",
          "locacao_totalQuarto",
          "locacao_totalDesconto",
          "locacao_totalLocacao",
          "motel_id",
          "quarto_id",
          "pessoa_id",
          "pagamentoforma_id",
          "usuario_id",
          "locacaoPosicao_id",
          "locacaoTipo_id"
        ]
      },
      "UpdateLocacaoDto": {
        "type": "object",
        "properties": {
          "locacao_totalItens": {
            "type": "number",
            "example": 129
          },
          "locacao_totalQuarto": {
            "type": "number",
            "example": 240
          },
          "locacao_totalDesconto": {
            "type": "number",
            "example": 20
          },
          "locacao_totalLocacao": {
            "type": "number",
            "example": 349
          },
          "motel_id": {
            "type": "number",
            "example": 1
          },
          "quarto_id": {
            "type": "number",
            "example": 10
          },
          "pessoa_id": {
            "type": "number",
            "example": 5
          },
          "pagamentoforma_id": {
            "type": "number",
            "example": 2
          },
          "usuario_id": {
            "type": "number",
            "example": 3
          },
          "locacaoPosicao_id": {
            "type": "number",
            "example": 1
          },
          "locacaoTipo_id": {
            "type": "number",
            "example": 1
          }
        }
      },
...
      "UpdatePessoaDto": {
        "type": "object",
        "properties": {
          "pessoa_nome": {
            "type": "string",
            "example": "Maria da Silva"
          },
          "pessoa_cpf": {
            "type": "string",
            "example": "12345678901"
          },
          "pessoa_telefone": {
            "type": "string",
            "example": "44999998888"
          },
          "pessoatipo_id": {
            "type": "number",
            "example": 1
          }
        }
      },
      "CreatePessoaTipoDto": {
        "type": "object",
        "properties": {
          "pessoatipo_descricao": {
            "type": "string",
            "example": "Funcionário",
            "description": "Descrição do tipo de pessoa."
          }
        },
        "required": [
          "pessoatipo_descricao"
        ]
      },
      "UpdatePessoaTipoDto": {
        "type": "object",
        "properties": {
          "pessoatipo_descricao": {
            "type": "string",
            "example": "Funcionário",
            "description": "Descrição do tipo de pessoa."
          }
        }
      },
...
      "CreateRegistroPontoDto": {
        "type": "object",
        "properties": {
          "usuario_id": {
            "type": "number",
            "example": 1,
            "description": "ID do usuário"
          },
          "motel_id": {
            "type": "number",
            "example": 10,
            "description": "ID do motel"
          },
          "registroponto_entrada": {
            "type": "boolean",
            "example": true,
            "description": "Define o tipo do registro"
          }
        },
        "required": [
          "usuario_id",
          "registroponto_entrada"
        ]
      },
      "UpdateRegistroPontoDto": {
        "type": "object",
        "properties": {
          "usuario_id": {
            "type": "number",
            "example": 1,
            "description": "ID do usuário"
          },
          "motel_id": {
            "type": "number",
            "example": 10,
            "description": "ID do motel"
          },
          "registroponto_entrada": {
            "type": "boolean",
            "example": true,
            "description": "Define o tipo do registro"
          }
        }
      },
...
    }
  }
}